server/requirements.txt
fastapi
psycopg2-binary
pydantic
python-dotenv
SQLAlchemy
starlette
typing
uvicorn